requirements.txt
flask
flask-cors
Pillow
RPi.GPIO
numpy
imutils
opencv-python
pyyaml